refactor: move contract-interfaces.ts
