+++ src/common/contract-interfaces.ts
