+++ src/common/index.ts
@@ -1,1 +1,2 @@
 export * from "./constants";
+export * from "./contract-interfaces";
