+++ test/TradeTrustERC721.test.ts
@@ -6,8 +6,7 @@ import { expect } from ".";
 import { deployTokenFixture } from "./fixtures";
 import { getTestUsers, getTitleEscrowContract, impersonateAccount, TestUsers } from "./utils";
 import { computeInterfaceId } from "./utils/computeInterfaceId";
-import { ContractInterfaces } from "./fixtures/contract-interfaces.fixture";
-import { AddressConstants } from "../src/common";
+import { ContractInterfaces, AddressConstants } from "../src/common";
 import { computeTitleEscrowAddress } from "../src/utils";
 
 const { loadFixture } = waffle;
